Answer the second review round on scan scope and named contracts
Scan scripts/cpd-renamed/run.ts with the ordinary 19-token scan and
ignore only the registry data file. Give the two provider slug rules
an explicit named contract. Split the config test suite into focused
files, keep the environment gate claims with the suite that already
owned them, and move the new builder and URL-seeded secure mode claims
there.
