+++ src/shared/config.ts
@@ -264,6 +264,9 @@ export const getTursoGroup = (): string => requireEnv("TURSO_GROUP");
 export const slugifyForProvider = (name: string, maxLength: number): string =>
   slugify(name).slice(0, maxLength).replace(/-+$/, "");
 
+/** Turns a site name into a provider-safe slug. */
+type SlugRule = (name: string) => string;
+
 /** A provider's resource-name rule: the provider slug capped at `maxLength`,
  * padded with `pad` when it lands under `minLength` (a provider that only
  * needs a non-empty name sets `minLength` to 1 and reads `pad` as the
@@ -277,7 +280,7 @@ const providerSlugRule =
     maxLength: number;
     minLength: number;
     pad: string;
-  }): ((name: string) => string) =>
+  }): SlugRule =>
   (name) => {
     const slug = slugifyForProvider(name, maxLength);
     return slug.length >= minLength
@@ -286,14 +289,14 @@ const providerSlugRule =
   };
 
 /** Deno Deploy app names: 3–32 chars, lowercase letters, numbers, hyphens. */
-export const denoDeployAppSlug = providerSlugRule({
+export const denoDeployAppSlug: SlugRule = providerSlugRule({
   maxLength: 32,
   minLength: 3,
   pad: "app",
 });
 
 /** Turso database names: non-empty, at most 63 chars. */
-export const tursoDatabaseSlug = providerSlugRule({
+export const tursoDatabaseSlug: SlugRule = providerSlugRule({
   maxLength: 63,
   minLength: 1,
   pad: "db",
